tweak for frontend use

+++ src/controllers/blogControllers.ts
@@ -34,8 +34,6 @@ export const get_blog = async (
 		} else {
 			res.status(200).json(blog)
 		}
-
-		// res.status(200).json(blog)
 	} catch (error) {
 		res.status(404).json({ error })
 	}
@@ -74,14 +72,6 @@ export const patch_blog = async (
 	} catch (error) {
 		res.status(404).json({ error })
 	}
-
-	// const blog = await Blog.findOneAndUpdate({ _id: id }, req.body)
-
-	// if (!blog) {
-	// 	return res.status(404).json({ error: 'No blog found' })
-	// }
-
-	// res.status(200).json(blog)
 }
 
 // DELETE Requests
@@ -92,9 +82,9 @@ export const delete_blog = async (
 ): Promise<void | Response> => {
 	const { id } = req.params
 
-	// if (!Types.ObjectId.isValid(id)) {
-	// 	return res.status(404).json({ error: 'No blog found' })
-	// }
+	if (!Types.ObjectId.isValid(id)) {
+		return res.status(404).json({ error: 'No blog found' })
+	}
 
 	try {
 		const blog = await Blog.findOneAndDelete({ _id: id })
@@ -104,8 +94,6 @@ export const delete_blog = async (
 		} else {
 			res.status(200).json(blog)
 		}
-
-		// res.status(200).json(blog)
 	} catch (error) {
 		res.status(404).json({ error })
 	}
